+++ src/ZeppelinForms/Core/Text/TextDocument.cs
@@ -1,0 +1,9 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace ZeppelinForms.Core.Text;
+
+internal class TextDocument
+{
+}